chore: more refactoring i18n
