+++ i18n/data/de.json
@@ -15,11 +15,6 @@
   "UIKit.LocalizedMoneyInput.hide": "Währungen verbergen ({remainingCurrencies})",
   "UIKit.LocalizedMoneyInput.missingRequiredField": "Pflichtfeld*, Geben Sie mindestens einen Wert an",
   "UIKit.LocalizedMoneyInput.show": "Alle Währungen anzeigen ({remainingCurrencies})",
-  "UIKit.LocalizedMultilineTextInput.collapse": "Verbergen",
-  "UIKit.LocalizedMultilineTextInput.expand": "Anzeigen",
-  "UIKit.LocalizedMultilineTextInput.hide": "Sprachen verbergen",
-  "UIKit.LocalizedMultilineTextInput.missingRequiredField": "Pflichtfeld*, Geben Sie mindestens einen Wert an",
-  "UIKit.LocalizedMultilineTextInput.show": "Alle Sprachen anzeigen",
   "UIKit.LocalizedRichTextInput.collapse": "Verbergen",
   "UIKit.LocalizedRichTextInput.expand": "Anzeigen",
   "UIKit.LocalizedRichTextInput.hide": "Sprachen verbergen",
