+++ i18n/data/en.json
@@ -15,11 +15,6 @@
   "UIKit.LocalizedMoneyInput.hide": "Hide currencies ({remainingCurrencies})",
   "UIKit.LocalizedMoneyInput.missingRequiredField": "This field is required. Provide at least one value.",
   "UIKit.LocalizedMoneyInput.show": "Show all currencies ({remainingCurrencies})",
-  "UIKit.LocalizedMultilineTextInput.collapse": "Collapse",
-  "UIKit.LocalizedMultilineTextInput.expand": "Expand",
-  "UIKit.LocalizedMultilineTextInput.hide": "Hide languages ({remainingLanguages})",
-  "UIKit.LocalizedMultilineTextInput.missingRequiredField": "This field is required. Provide at least one value.",
-  "UIKit.LocalizedMultilineTextInput.show": "Show all languages ({remainingLanguages})",
   "UIKit.LocalizedRichTextInput.collapse": "Collapse",
   "UIKit.LocalizedRichTextInput.expand": "Expand",
   "UIKit.LocalizedRichTextInput.hide": "Hide languages ({remainingLanguages})",
