+++ i18n/data/es.json
@@ -15,11 +15,6 @@
   "UIKit.LocalizedMoneyInput.hide": "Ocultar divisas ({remainingCurrencies})",
   "UIKit.LocalizedMoneyInput.missingRequiredField": "Este campo es obligatorio. Proporcione un valor al menos.",
   "UIKit.LocalizedMoneyInput.show": "Mostrar todas las divisas ({remainingCurrencies})",
-  "UIKit.LocalizedMultilineTextInput.collapse": "Contraer",
-  "UIKit.LocalizedMultilineTextInput.expand": "Expandir",
-  "UIKit.LocalizedMultilineTextInput.hide": "Ocultar lenguajes ({remainingLanguages})",
-  "UIKit.LocalizedMultilineTextInput.missingRequiredField": "Este campo es obligatorio. Proporcione un valor al menos.",
-  "UIKit.LocalizedMultilineTextInput.show": "Mostrar todos los lenguajes ({remainingLanguages})",
   "UIKit.LocalizedRichTextInput.collapse": "",
   "UIKit.LocalizedRichTextInput.expand": "",
   "UIKit.LocalizedRichTextInput.hide": "",
