+++ i18n/data/fr-FR.json
@@ -15,11 +15,6 @@
   "UIKit.LocalizedMoneyInput.hide": "Masquer les devices ({remainingCurrencies})",
   "UIKit.LocalizedMoneyInput.missingRequiredField": "Ce champ est requis. Veuillez indiquer une valeur.",
   "UIKit.LocalizedMoneyInput.show": "Afficher toutes les devises  ({remainingCurrencies})",
-  "UIKit.LocalizedMultilineTextInput.collapse": "Réduire",
-  "UIKit.LocalizedMultilineTextInput.expand": "Développer",
-  "UIKit.LocalizedMultilineTextInput.hide": "Masquer les langues  ({remainingLanguages})",
-  "UIKit.LocalizedMultilineTextInput.missingRequiredField": "Ce champ est requis. Veuillez indiquer au moins une valeur.",
-  "UIKit.LocalizedMultilineTextInput.show": "Afficher toutes les langues ({remainingLanguages})",
   "UIKit.LocalizedRichTextInput.collapse": "",
   "UIKit.LocalizedRichTextInput.expand": "",
   "UIKit.LocalizedRichTextInput.hide": "",
