+++ i18n/data/zh-CN.json
@@ -15,11 +15,6 @@
   "UIKit.LocalizedMoneyInput.hide": "隐藏货币 ({remainingCurrencies})",
   "UIKit.LocalizedMoneyInput.missingRequiredField": "此字段为必填项。请提供至少一个值。",
   "UIKit.LocalizedMoneyInput.show": "显示所有货币 ({remainingCurrencies})",
-  "UIKit.LocalizedMultilineTextInput.collapse": "崩溃",
-  "UIKit.LocalizedMultilineTextInput.expand": "扩展",
-  "UIKit.LocalizedMultilineTextInput.hide": "隐藏语言 ({remainingLanguages})",
-  "UIKit.LocalizedMultilineTextInput.missingRequiredField": "此字段为必填项。请提供至少一个值。",
-  "UIKit.LocalizedMultilineTextInput.show": "显示所有语言 ({remainingLanguages})",
   "UIKit.LocalizedRichTextInput.collapse": "",
   "UIKit.LocalizedRichTextInput.expand": "",
   "UIKit.LocalizedRichTextInput.hide": "",
