--- src/components/inputs/localized-multiline-text-input/messages.js
@@ -1,29 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  missingRequiredField: {
-    id: 'UIKit.LocalizedMultilineTextInput.missingRequiredField',
-    description: 'Error message for missing required value',
-    defaultMessage: 'This field is required. Provide at least one value.',
-  },
-  show: {
-    id: 'UIKit.LocalizedMultilineTextInput.show',
-    description: '',
-    defaultMessage: 'Show all languages ({remainingLanguages})',
-  },
-  hide: {
-    id: 'UIKit.LocalizedMultilineTextInput.hide',
-    description: '',
-    defaultMessage: 'Hide languages ({remainingLanguages})',
-  },
-  expand: {
-    id: 'UIKit.LocalizedMultilineTextInput.expand',
-    description: '',
-    defaultMessage: 'Expand',
-  },
-  collapse: {
-    id: 'UIKit.LocalizedMultilineTextInput.collapse',
-    description: '',
-    defaultMessage: 'Collapse',
-  },
-});
