+++ src/components/inputs/localized-multiline-text-input/required-value-error-message.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import ErrorMessage from '../../messages/error-message';
-import messages from './messages';
+import messages from '../../internals/messages/localized-input';
 
 const RequiredValueErrorMessage = () => (
   <ErrorMessage>
